ENH: Allow -s to be an optional argument
The current version of the code takes a pattern and a list of subjects.
Each item in the list of subjects is substituted into the pattern to
create a glob pattern or a path for a directory. The output of this
substitution is used to find all the dicoms for the subject.

Currently there is no way to skip the substitution of the subject name,
as a list of subject names is a required argument. However, some bits
of the code seems to be crafted to allow the subject list to be
optional.

This makes the subject list an optional argument.